+++ app/src/main/java/com/mateusz113/financemanager/domain/model/PaymentListing.kt
@@ -5,7 +5,7 @@ import java.time.LocalDate
 data class PaymentListing(
     val id: String,
     val title: String,
-    val amount: Float,
+    val amount: Double,
     val date: LocalDate,
     val category: Category
 )
